+++ src/Command/TorrentList.php
@@ -19,6 +19,7 @@ class TorrentList extends Command
             ->setAliases(['tl'])
             ->setDescription('List torrents')
             ->addOption('sort', null, InputOption::VALUE_OPTIONAL, 'Sort by column number', 4)
+            ->addOption('age', null, InputOption::VALUE_OPTIONAL, 'Sort by torrent age, ex. \'>1 <5\'')
             ->setHelp(<<<EOT
 The <info>torrent-list</info> list torrents.
 EOT
@@ -32,22 +33,20 @@ EOT
         $client = $this->getApplication()->getClient();
 
         $torrentList = $client->getTorrentData();
-        if (empty($torrentList)) {
-            return 0;
-        }
+
+        $torrentList = $client->filterTorrents($torrentList, [
+            'age' => $input->getOption('age'),
+        ]);
 
         $rows = [];
         foreach ($torrentList as $torrent) {
-            $ageDays = $perDay = '-';
-            if ($torrent[Torrent\Get::DONE_DATE]) {
-                $age = time() - $torrent[Torrent\Get::DONE_DATE];
-                $perDay = round($torrent[Torrent\Get::UPLOAD_EVER] / $age * 86400 / 1024 / 1000 / 1000, 2);
-                $ageDays = round($age / 86400, 0);
-            }
+            $age = $client->getTorrentAgeInDays($torrent);
+            $perDay = $age ? round($torrent[Torrent\Get::UPLOAD_EVER] / $age / 1024 / 1000 / 1000, 2) : 0;
+
             $rows[] = [
                 $torrent[Torrent\Get::NAME],
                 $torrent[Torrent\Get::ID],
-                $ageDays,
+                $age,
                 round($torrent[Torrent\Get::DOWNLOAD_EVER] / 1024 / 1000 / 1000, 2),
                 round($torrent[Torrent\Get::UPLOAD_EVER] / 1024 / 1000 / 1000, 2),
                 $perDay,
@@ -59,10 +58,11 @@ EOT
             '',
             '',
             round($client->getTorrentsSize($torrentList) / 1024 / 1000 / 1000, 2),
-            round($client->getTorrentsSize($torrentList, Torrent\Get::UPLOAD_EVER) / 1024 / 1000 / 1000, 2)
+            round($client->getTorrentsSize($torrentList, Torrent\Get::UPLOAD_EVER) / 1024 / 1000 / 1000, 2),
+            ''
         ];
 
-        $this->sortTable($rows, $input);
+        $rows = $this->sortTable($rows, $input);
 
         $table = new Table($output);
         $table->setHeaders($headers);
@@ -70,21 +70,22 @@ EOT
         $table->addRow(new TableSeparator());
         $table->addRow($totals);
         $table->render();
-
-        return 0;
     }
 
     private function sortTable(array $rows, InputInterface $input)
     {
-        $columnsTotal = count($rows[0]);
+        $rowsSorted = $rows;
+        $columnsTotal = count(end($rows));
 
         $sortColumn = max(1, min(
             $columnsTotal,
             $input->getOption('sort')
         )) - 1;
 
-        usort($rows, function ($first, $second) use ($sortColumn) {
+        usort($rowsSorted, function ($first, $second) use ($sortColumn) {
             return $first[$sortColumn] > $second[$sortColumn] ? 1 : -1;
         });
+
+        return $rowsSorted;
     }
 }