+++ CHANGELOG.md
@@ -4,7 +4,7 @@ BREAKING CHANGES:
 
 FEATURES:
   * packagist.org package: now able install with composer
-  * torrent-list: add columns 'Age', 'Uploaded', 'Per day', add --sort option
+  * torrent-list: add columns 'Age', 'Uploaded', 'Per day', add --sort, --age options
 
 IMPROVEMENTS: